Turn deleting old cache data in Redis into generic migration
