+++ cmd/frontend/internal/cli/serve_cmd.go
@@ -151,6 +151,7 @@ func Main() error {
 	goroutine.Go(func() { bg.MigrateAllSettingsMOTDToNotices(context.Background()) })
 	goroutine.Go(func() { bg.MigrateSavedQueriesAndSlackWebhookURLsFromSettingsToDatabase(context.Background()) })
 	goroutine.Go(func() { bg.LogSearchQueries(context.Background()) })
+	goroutine.Go(func() { bg.DeleteOldCacheDataInRedis() })
 	goroutine.Go(mailreply.StartWorker)
 	go updatecheck.Start()
 	if hooks.AfterDBInit != nil {
